use go-tuf instead of go-tuf-metadata
