--- clients/go-tuf-metadata/.gitignore
@@ -1,1 +1,0 @@
-/go-tuf-metadata
