+++ clients/go-tuf/.gitignore
@@ -1,0 +1,1 @@
+/go-tuf-metadata
