+++ clients/go-tuf/go.mod
@@ -1,0 +1,31 @@
+module github.com/jku/tuf-conformance/clients/go-tuf
+
+go 1.21
+
+toolchain go1.22.3
+
+require (
+	github.com/jku/tuf-conformance/clients/go-tuf-metadata v0.0.0-20240430071659-3e958fed3625
+	github.com/sirupsen/logrus v1.9.3
+	github.com/spf13/cobra v1.8.1
+	github.com/theupdateframework/go-tuf/v2 v2.0.0-20240617113600-41866142c66a
+)
+
+require (
+	github.com/google/go-containerregistry v0.19.1 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/kr/text v0.1.0 // indirect
+	github.com/letsencrypt/boulder v0.0.0-20230907030200-6d76a0f91e1e // indirect
+	github.com/opencontainers/go-digest v1.0.0 // indirect
+	github.com/rdimitrov/go-tuf-metadata v0.0.0-20230710113934-fe6fba86596c // indirect
+	github.com/secure-systems-lab/go-securesystemslib v0.8.0 // indirect
+	github.com/sigstore/sigstore v1.8.4 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/titanous/rocacheck v0.0.0-20171023193734-afe73141d399 // indirect
+	golang.org/x/crypto v0.23.0 // indirect
+	golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
+	golang.org/x/sys v0.20.0 // indirect
+	golang.org/x/term v0.20.0 // indirect
+	gopkg.in/go-jose/go-jose.v2 v2.6.3 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
